feat(fe): make confirmnavigation reusable
