+++ apps/frontend/app/admin/contest/create/page.tsx
@@ -31,11 +31,12 @@ import { zodResolver } from '@hookform/resolvers/zod'
 import { PlusCircleIcon } from 'lucide-react'
 import Link from 'next/link'
 import { useRouter } from 'next/navigation'
-import { useEffect, useState, useRef } from 'react'
+import { useState, useRef } from 'react'
 import { useForm, FormProvider } from 'react-hook-form'
 import { FaAngleLeft } from 'react-icons/fa6'
 import { IoMdCheckmarkCircleOutline } from 'react-icons/io'
 import { toast } from 'sonner'
+import { useConfirmNavigation } from '../../_components/ConfirmNavigation'
 import DescriptionForm from '../../_components/DescriptionForm'
 import FormSection from '../../_components/FormSection'
 import SwitchField from '../../_components/SwitchField'
@@ -55,39 +56,7 @@ export default function Page() {
   const shouldSkipWarning = useRef(false)
   const router = useRouter()
 
-  const useConfirmNavigation = () => {
-    const router = useRouter()
-    const handleBeforeUnload = (event: BeforeUnloadEvent) => {
-      event.preventDefault()
-      event.returnValue = ''
-    }
-
-    useEffect(() => {
-      window.addEventListener('beforeunload', handleBeforeUnload)
-
-      const originalPush = router.push
-
-      router.push = (href, ...args) => {
-        if (shouldSkipWarning.current) {
-          originalPush(href, ...args)
-          return
-        }
-        const shouldWarn = window.confirm(
-          'Are you sure you want to leave this page? Changes you made may not be saved.'
-        )
-        if (shouldWarn) {
-          originalPush(href, ...args)
-        }
-      }
-
-      return () => {
-        window.removeEventListener('beforeunload', handleBeforeUnload)
-        router.push = originalPush
-      }
-    }, [router])
-  }
-
-  useConfirmNavigation()
+  useConfirmNavigation(shouldSkipWarning.current)
 
   const methods = useForm<CreateContestInput>({
     resolver: zodResolver(createSchema),
