+++ apps/frontend/app/admin/problem/[id]/edit/page.tsx
@@ -1,5 +1,6 @@
 'use client'
 
+import { useConfirmNavigation } from '@/app/admin/_components/ConfirmNavigation'
 import { Button } from '@/components/ui/button'
 import { ScrollArea, ScrollBar } from '@/components/ui/scroll-area'
 import { UPDATE_PROBLEM } from '@/graphql/problem/mutations'
@@ -9,7 +10,7 @@ import type { Template, Testcase, UpdateProblemInput } from '@generated/graphql'
 import { zodResolver } from '@hookform/resolvers/zod'
 import Link from 'next/link'
 import { useRouter } from 'next/navigation'
-import { useEffect, useRef, useState } from 'react'
+import { useRef, useState } from 'react'
 import { useForm, FormProvider } from 'react-hook-form'
 import { FaAngleLeft } from 'react-icons/fa6'
 import { IoIosCheckmarkCircle } from 'react-icons/io'
@@ -33,39 +34,7 @@ export default function Page({ params }: { params: { id: string } }) {
   const shouldSkipWarning = useRef(false)
   const router = useRouter()
 
-  const useConfirmNavigation = () => {
-    const router = useRouter()
-    const handleBeforeUnload = (event: BeforeUnloadEvent) => {
-      event.preventDefault()
-      event.returnValue = ''
-    }
-
-    useEffect(() => {
-      window.addEventListener('beforeunload', handleBeforeUnload)
-
-      const originalPush = router.push
-
-      router.push = (href, ...args) => {
-        if (shouldSkipWarning.current) {
-          originalPush(href, ...args)
-          return
-        }
-        const shouldWarn = window.confirm(
-          'Are you sure you want to leave this page? Changes you made may not be saved.'
-        )
-        if (shouldWarn) {
-          originalPush(href, ...args)
-        }
-      }
-
-      return () => {
-        window.removeEventListener('beforeunload', handleBeforeUnload)
-        router.push = originalPush
-      }
-    }, [router])
-  }
-
-  useConfirmNavigation()
+  useConfirmNavigation(shouldSkipWarning.current)
 
   const methods = useForm<UpdateProblemInput>({
     resolver: zodResolver(editSchema),
